backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
sqlalchemy
aiofiles
opencv-python-headless
torch
torchvision
Pillow
python-dotenv
sqlalchemy
fastapi-cors
passlib[bcrypt]
python-jose[cryptography]